requirements.txt
aiogram==2.25.1
asyncpg
python-dotenv
